# apps/backend/src/common/dto/paginated.dto.ts
import { ApiProperty, ApiExtraModels } from '@nestjs/swagger';

@ApiExtraModels()
export class PaginatedDto<T> {
  @ApiProperty({ description: 'Položky výsledku', isArray: true })
  items: T[];

  @ApiProperty({ example: 1 })
  page: number;

  @ApiProperty({ example: 10 })
  limit: number;

  @ApiProperty({ example: 25 })
  total: number;
}